Use an explicit cast to avoid a warning on 64-bit systems
git-svn-id: https://svn.nmap.org@29979 e0a8ed71-7df4-0310-8962-fdc924857419

+++ nping/utils.cc
@@ -469,7 +469,7 @@ const char *u64tostr(u64 number){
   #ifdef WIN32
     snprintf(mynumber, 32, "%I64u", number);
   #else
-    snprintf(mynumber, 32, "%llu", number);
+    snprintf(mynumber, 32, "%llu", (long long unsigned int) number);
   #endif
   return mynumber;
 } /* End of u64tostr() */
